+++ isla-lib/Cargo.toml
@@ -23,10 +23,14 @@ lalrpop-util = "0.19.0"
 crossbeam = "0.7.3"
 lazy_static = "1.4.0"
 toml = "0.5.5"
-z3-sys = "*"
+z3-sys = { path = "C:\\Users\\Benni\\repositories\\z3.rs\\z3-sys" }
 libc = "0.2.5"
 serde = "1.0.104"
 bincode = "1.2.1"
 sha2 = "0.8.1"
 petgraph = "0.5.0"
-elfloader = "0.10"
+elfloader = { path = "C:\\Users\\Benni\\repositories\\rust-elfloader" }
+
+[[bin]]
+name = "test"
+path = "src/test.rs"
