+++ isla-lib/src/lib.rs
@@ -56,3 +56,4 @@ pub mod simplify;
 pub mod smt;
 pub mod type_check;
 pub mod zencode;
+pub mod elf_loader;
